Rename "MenuItem" to "LegacyMenuItem" and "NewMenuItem" to "MenuItem"
