+++ src/docs/components/DemoApp.tsx
@@ -18,7 +18,7 @@ import HeaderTitle from "../../components/HeaderTitle"
 import Intro from "../../components/Intro"
 import Layout from "../../components/Layout"
 import HashRouter from "../../components/HashRouter"
-import MenuItem from "../../components/MenuItem"
+import MenuItem from "../../components/LegacyMenuItem"
 import Page from "../../components/Page"
 import Search from "../../components/Search"
 import SidebarMenu from "../../components/SidebarMenu"
