+++ src/docs/examples/IconMenuItem.tsx
@@ -1,4 +1,4 @@
-import MenuItem from "../../components/MenuItem"
+import MenuItem from "../../components/LegacyMenuItem"
 import SidebarMenu from "../../components/SidebarMenu"
 import IconMenuItem from "../../components/IconMenuItem"
 import { BoxArrowLeft, Gear, List, Person } from "react-bootstrap-icons"
@@ -18,7 +18,12 @@ export default function IconMenuItemExample() {
                             icon={<Person className="bi" />}
                         />
 
-                        <IconMenuItem href="#" outerClass="flex-fill" label="Settings" icon={<Gear className="bi" />} />
+                        <IconMenuItem
+                            href="#"
+                            outerClass="flex-fill"
+                            label="Settings"
+                            icon={<Gear className="bi" />}
+                        />
 
                         <IconMenuItem
                             href="#"
@@ -34,13 +39,17 @@ export default function IconMenuItemExample() {
                     label="Toggle menu"
                     onClick={() => {
                         window.blueHashRouterRef.setState({
-                            expandSidebar: !window.blueHashRouterRef.state.expandSidebar
+                            expandSidebar:
+                                !window.blueHashRouterRef.state.expandSidebar
                         })
                     }}
                 />
 
                 <MenuItem icon={<span>😅</span>} label="Hello World" />
-                <MenuItem icon={<span style={{ transform: "scale(-1, 1)" }}>👀</span>} label="Another normal item" />
+                <MenuItem
+                    icon={<span style={{ transform: "scale(-1, 1)" }}>👀</span>}
+                    label="Another normal item"
+                />
             </SidebarMenu>
         </div>
     )
