+++ src/docs/examples/LegacyMenuItem/Sidebar.tsx
@@ -1,10 +1,13 @@
-import MenuItem from "../../../components/MenuItem"
+import MenuItem from "../../../components/LegacyMenuItem"
 import SidebarMenu from "../../../components/SidebarMenu"
 
 export default function MenuItemExample() {
     return (
         <div className="bg-primary position-relative">
-            <SidebarMenu sidebarClass="position-static shadow-none overflow-visible" menuClass="overflow-visible">
+            <SidebarMenu
+                sidebarClass="position-static shadow-none overflow-visible"
+                menuClass="overflow-visible"
+            >
                 <MenuItem icon={<span>😅</span>} label="Hello World" />
                 <MenuItem
                     icon={<span style={{ transform: "scale(-1, 1)" }}>👀</span>}
@@ -17,7 +20,9 @@ export default function MenuItemExample() {
                     </MenuItem>
                 </MenuItem>
             </SidebarMenu>
-            <p className="p-3 bg-body">Resize the browser to see how the sidebar and its items behave.</p>
+            <p className="p-3 bg-body">
+                Resize the browser to see how the sidebar and its items behave.
+            </p>
         </div>
     )
 }
