+++ src/docs/examples/MenuItem/Behave as collapse group.tsx
@@ -1,15 +1,15 @@
 import { DropletFill } from "react-bootstrap-icons"
-import NewMenuItem from "../../../components/NewMenuItem"
+import MenuItem from "../../../components/MenuItem"
 
 export default function AsCollapseGroupExample() {
     return (
-        <NewMenuItem
+        <MenuItem
             iconBefore={<DropletFill />}
             label="Parent"
             as="collapse-group"
         >
-            <NewMenuItem label="Child 1" />
-            <NewMenuItem label="Child 2" />
-        </NewMenuItem>
+            <MenuItem label="Child 1" />
+            <MenuItem label="Child 2" />
+        </MenuItem>
     )
 }
