+++ src/docs/examples/MenuItem/Is active.tsx
@@ -1,0 +1,5 @@
+import MenuItem from "../../../components/MenuItem"
+
+export default function IsActiveExample() {
+    return <MenuItem label="I am active" active />
+}
