--- src/docs/examples/NewMenuItem/Is active.tsx
@@ -1,5 +1,0 @@
-import NewMenuItem from "../../../components/NewMenuItem"
-
-export default function IsActiveExample() {
-    return <NewMenuItem label="I am active" active />
-}
